code/requirements.txt
easydict==1.10
matplotlib==3.5.1
numpy==1.22.4
pandas==1.4.3
PyPDF2==2.11.2
PyYAML==6.0
scipy==1.7.3
seaborn==0.12.1
torch==1.13.0
torchvision==0.14.0
